Bug 58897 : Step 5 misc clean up remove empty test remove commented code split some tests explicitly disable org.apache.jmeter.protocol.http.sampler.PackageTest when in headless mode
