+++ test/src/org/apache/jmeter/functions/RandomFunctionTest.java
@@ -1,0 +1,48 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * 
+ */
+
+package org.apache.jmeter.functions;
+
+import static org.apache.jmeter.functions.FunctionTestHelper.makeParams;
+
+import java.util.Collection;
+
+import org.apache.jmeter.engine.util.CompoundVariable;
+import org.apache.jmeter.junit.JMeterTestCase;
+import org.junit.Test;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertEquals;
+
+public class RandomFunctionTest extends JMeterTestCase {
+
+    @Test
+    public void randomTest1() throws Exception {
+        Random r = new Random();
+        Collection<CompoundVariable> parms = makeParams("0","10000000000","VAR");
+        r.setParameters(parms);
+        String s = r.execute(null,null);
+        long l = Long.parseLong(s);
+        assertTrue(l>=0 && l<=10000000000L);
+        
+        parms = makeParams("1","1","VAR");
+        r.setParameters(parms);
+        s = r.execute(null,null);
+        l = Long.parseLong(s);
+        assertEquals(1, l);
+    }
+}
