+++ test/src/org/apache/jmeter/testbeans/gui/PackageTest.java
@@ -35,7 +35,6 @@ import org.apache.jorphan.reflect.ClassFinder;
 import org.apache.log.Logger;
 
 import junit.framework.Test;
-// import junit.framework.TestCase;
 import junit.framework.TestSuite;
 
 /*
@@ -50,9 +49,7 @@ import junit.framework.TestSuite;
 public final class PackageTest extends JMeterTestCaseJUnit3 {
     private static final Logger log = LoggingManager.getLoggerForClass();
 
-    // ResourceBundle i18nEdit=
-    // ResourceBundle.getBundle("org.apache.jmeter.resources.i18nedit");
-    private static final Locale defaultLocale = new Locale("en",""); // i18nEdit.getString("locale.default");
+    private static final Locale defaultLocale = new Locale("en","");
 
     private final ResourceBundle defaultBundle;
 
@@ -81,7 +78,7 @@ public final class PackageTest extends JMeterTestCaseJUnit3 {
     @Override
     public void setUp() {
         if (testLocale == null) {
-            return;// errorDetected()
+            return;
         }
         JMeterUtils.setLocale(testLocale);
         Introspector.flushFromCaches(testBeanClass);
@@ -106,7 +103,7 @@ public final class PackageTest extends JMeterTestCaseJUnit3 {
     public void runTest() throws Throwable {
         if (testLocale == null) {
             super.runTest();
-            return;// errorDetected()
+            return;
         }
         if (bundle == defaultBundle) {
             checkAllNecessaryKeysPresent();
@@ -155,7 +152,6 @@ public final class PackageTest extends JMeterTestCaseJUnit3 {
             String name = descriptors[i].getName();
 
             bundle.getString(name + ".displayName");
-            // bundle.getString(name+".shortDescription"); NOT MANDATORY
 
             String group = (String) descriptors[i].getValue(GenericTestBeanCustomizer.GROUP);
             if (group != null) {
@@ -179,8 +175,7 @@ public final class PackageTest extends JMeterTestCaseJUnit3 {
                         GenericTestBeanCustomizer.RESOURCE_BUNDLE);
             } catch (IntrospectionException e) {
                 log.error("Can't get beanInfo for " + testBeanClass.getName(), e);
-                throw new Error(e.toString(), e); // Programming error. Don't
-                                                // continue.
+                throw new Error(e.toString(), e); // Programming error. Don't continue.
             }
 
             if (defaultBundle == null) {
@@ -190,15 +185,14 @@ public final class PackageTest extends JMeterTestCaseJUnit3 {
                 }
                 errorDetected=true;
                 log.error("No default bundle found for " + className + " using " + defaultLocale.toString());
-                //throw new Error("No default bundle for class " + className);
                 continue;
             }
 
             suite.addTest(new PackageTest(testBeanClass, defaultLocale, defaultBundle));
 
-            String [] languages = JMeterMenuBar.getLanguages();
-            for (int i=0; i < languages.length; i++){
-                final String[] language = languages[i].split("_");
+            String[] languages = JMeterMenuBar.getLanguages();
+            for (String lang : languages) {
+                final String[] language = lang.split("_");
                 if (language.length == 1){
                     suite.addTest(new PackageTest(testBeanClass, new Locale(language[0]), defaultBundle));                                    
                 } else if (language.length == 2){
